coll: Optimize datatype validation to int16
Optimize the transparent datatype validation mechanism to use int16
instead of int64 for the allreduce operation, reducing message size
from 32 bytes to 8 bytes (75% reduction).

Rationale:
At the OMPI layer, only UCC predefined datatypes are supported for
collectives. Generic (user-defined) datatypes are not used, which
allows us to make assumptions about the value range:

- Predefined datatype maximum value: 136 (UCC_DT_FLOAT128_COMPLEX = 17 << 3)
- Memory type maximum value: 5 (UCC_MEMORY_TYPE_LAST)
- Both values fit comfortably in int16 range (-32768 to 32767)

The optimization changes:
- int64_t arrays → int16_t arrays in ucc_dt_check_state_t
- UCC_DT_INT64 → UCC_DT_INT16 in service allreduce call
- Update all casts from (int64_t) to (int16_t)
- Message size: 32 bytes → 8 bytes per validation operation
- Added UCC_DT_IS_PREDEFINED() check to explicitly reject generic
  datatypes and prevent int16 overflow from user arguments

Benefits:
- 75% bandwidth reduction for validation messages
- Same O(1) message size (doesn't scale with number of ranks)
- Same correctness guarantees (min/max trick still works)
- No additional latency (still single allreduce)
- Safe against overflow from generic datatype pointer values

Note: This optimization is valid only for predefined datatypes.
Generic datatypes use full 64-bit pointer values and would differ
across processes anyway, causing validation to correctly reject them.

+++ src/schedule/ucc_schedule.h
@@ -80,14 +80,15 @@ typedef struct ucc_service_coll_req ucc_service_coll_req_t;
  *   - Send [dt, -dt, mem, -mem] with MIN reduction
  *   - After reduction: if min(dt) == -min(-dt), all ranks have same dt
  *   - Same principle for memory type
- *   - Message size: constant 4 integers (doesn't scale with number of ranks)
+ *   - Message size: 8 bytes (4 × int16, doesn't scale with number of ranks)
+ *   - Optimized for predefined datatypes only (max value 136 fits in int16)
  *
  * If validation fails, the dependency mechanism prevents the actual task from posting.
  */
 typedef struct ucc_dt_check_state {
     ucc_service_coll_req_t *check_req;        /* Service allreduce request */
-    int64_t                 reduced_values[4];/* Result: [min(dt), min(-dt), min(mem), min(-mem)] */
-    int64_t                 local_values[4];  /* Local: [dt, -dt, mem, -mem] */
+    int16_t                 reduced_values[4];/* Result: [min(dt), min(-dt), min(mem), min(-mem)] */
+    int16_t                 local_values[4];  /* Local: [dt, -dt, mem, -mem] */
     ucc_subset_t            subset;           /* Subset for service allreduce */
     int                     validated;        /* 1 if validation passed, 0 if failed */
     struct ucc_coll_task   *actual_task;      /* Pointer to actual collective task */
